refactor: keep console env helpers local
